+++ src/lapidary/runtime/model/op.py
@@ -1,8 +1,11 @@
 import inspect
 from collections.abc import Awaitable, Callable
 
+import httpx
 import typing_extensions as typing
 
+from ..middleware import HttpxMiddleware
+from ..types_ import Next
 from .error import HttpErrorResponse
 from .request import RequestAdapter, prepare_request_adapter
 from .response import ResponseMessageExtractor, mk_response_extractor
@@ -24,6 +27,22 @@ def process_operation_method(fn: Callable, op: 'Operation') -> tuple[RequestAdap
         raise TypeError(fn.__name__) from error
 
 
+def _wrap_middleware(mw_: HttpxMiddleware, next_: Next) -> Next:
+    async def wrapped(req: httpx.Request) -> httpx.Response:
+        return await mw_(req, next_)
+
+    return wrapped
+
+
+def _mk_send(client_send, auth: httpx.Auth) -> Next:
+    async def send(request: httpx.Request) -> httpx.Response:
+        resp = await client_send(request, auth=auth)
+        await resp.aread()
+        return resp
+
+    return send
+
+
 def mk_exchange_fn(
     op_method: Callable,
     op_decorator: 'Operation',
@@ -33,17 +52,11 @@ def mk_exchange_fn(
     async def exchange(self: 'ClientBase', **kwargs) -> typing.Any:
         request, auth = request_adapter.build_request(self, kwargs)
 
-        mw_state = []
-        for mw in self._middlewares:
-            mw_state.append(await mw.handle_request(request))
+        send = _mk_send(self._client.send, auth)
+        for middleware in reversed(self._middlewares):
+            send = _wrap_middleware(middleware, send)
 
-        response = await self._client.send(request, auth=auth)
-
-        await response.aread()
-
-        for mw, state in zip(reversed(self._middlewares), reversed(mw_state)):
-            await mw.handle_response(response, request, state)
-
+        response = await send(request)
         status_code, result = response_handler.handle_response(response)
         if status_code >= 400:
             raise HttpErrorResponse(status_code, result[1], result[0])