+++ ChangeLog.md
@@ -14,6 +14,7 @@ and the format of this file is based on [Keep a Changelog](https://keepachangelo
 
 - Bumped minimum required Python version to 3.10
 - ClientBase accepts httpx client instance; remove async context manager protocol
+- Change middleware interface to nestable callable
 
 
 ## [0.12.3] - 2025-03-01